+++ mmworkbench/components/role_classifier.py
@@ -12,6 +12,7 @@ from sklearn.externals import joblib
 
 from ..models import create_model, ENTITY_EXAMPLE_TYPE, CLASS_LABEL_TYPE
 from ..core import Query
+from ..constants import DEFAULT_TRAIN_SET_REGEX
 
 from .classifier import Classifier, ClassifierConfig
 from ._config import get_classifier_config
@@ -61,7 +62,7 @@ class RoleClassifier(Classifier):
                                               entity=self.entity_type)
         return super()._get_model_config(loaded_config, **kwargs)
 
-    def fit(self, queries=None, label_set='train', previous_model_path=None, **kwargs):
+    def fit(self, queries=None, label_set=None, previous_model_path=None, **kwargs):
         """Trains a statistical model for role classification using the provided training examples
 
         Args:
@@ -79,6 +80,11 @@ class RoleClassifier(Classifier):
         # create model with given params
         model_config = self._get_model_config(**kwargs)
         model = create_model(model_config)
+
+        if not label_set:
+            label_set = model_config.train_label_set
+            label_set = label_set if label_set else DEFAULT_TRAIN_SET_REGEX
+
         new_hash = self._get_model_hash(model_config, queries, label_set)
 
         if previous_model_path:
@@ -191,7 +197,7 @@ class RoleClassifier(Classifier):
         """
         raise NotImplementedError
 
-    def _get_query_tree(self, queries=None, label_set='train', raw=False):
+    def _get_query_tree(self, queries=None, label_set=DEFAULT_TRAIN_SET_REGEX, raw=False):
         """Returns the set of queries to train on
 
         Args:
@@ -211,7 +217,7 @@ class RoleClassifier(Classifier):
         return self._resource_loader.get_labeled_queries(domain=self.domain, intent=self.intent,
                                                          label_set=label_set, raw=raw)
 
-    def _get_queries_and_labels(self, queries=None, label_set='train'):
+    def _get_queries_and_labels(self, queries=None, label_set=DEFAULT_TRAIN_SET_REGEX):
         """Returns a set of queries and their labels based on the label set
 
         Args:
@@ -244,7 +250,7 @@ class RoleClassifier(Classifier):
 
         return examples, labels
 
-    def _get_queries_and_labels_hash(self, queries=None, label_set='train'):
+    def _get_queries_and_labels_hash(self, queries=None, label_set=DEFAULT_TRAIN_SET_REGEX):
         query_tree = self._get_query_tree(queries, label_set=label_set, raw=True)
         queries = self._resource_loader.flatten_query_tree(query_tree)
         queries.sort()